requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.2
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pdfkit==0.6.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sendgrid==6.4.7
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
